% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{bcf_getWindows}
\alias{bcf_getWindows}
\title{Generate a list of regions which slide over the genome}
\usage{
bcf_getWindows(path, windowsize = 1000, slide = 1000)
}
\arguments{
\item{path}{Path to VCF or BCF file}

\item{windowsize}{Size of each window, in bases}

\item{slide}{Step size of each window, in bases}
}
\description{
Generate a list of regions which slide over the genome
}
