% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deerData.R
\name{deerData}
\alias{deerData}
\title{Deer Dataset}
\usage{
deerData()
}
\value{
A character string of the complete path to the VCF file of the deer data set.
}
\description{
Function for extracting the path to the (VCF) file of the deer data set used in the publication
by \insertCite{bilton2018genetics2;textual}{GUSbase}.
}
\details{
The data consists of 38 SNPs and 704 deer samples genotyped using the genotyping-by-sequencing
method. The data is in VCF format (see this \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{page}
for specification of VCF format).
}
\examples{
## extract the name of the vcf file for the deer dataset
deerData()

}
\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
