% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSLD2.R
\name{GUS_LD}
\alias{GUS_LD}
\title{Estimate LD using the GUS-LD model}
\usage{
GUS_LD(genon, depth_Ref, depth_Alt, LDmeasure = c("Dprime", "r2"))
}
\arguments{
\item{genon}{012-coded genotype matrix. SNPs are columns.}

\item{depth_Ref}{Reference/major allele depth, same dims as `genon`}

\item{depth_Alt}{Alternate/minor allele depth, same dims as `genon`}

\item{LDmeasure}{Measure of LD. One or both of 'Dprime' and 'r2' (case sensitive)}
}
\value{
list with slots for each LDmeasure
}
\description{
Estimate LD using the GUS-LD model
}
