% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSLD2.R
\name{GUSLD_core}
\alias{GUSLD_core}
\title{Title}
\usage{
GUSLD_core(snp1, genon, depth_Ref, depth_Alt, freq, LDmeasures)
}
\arguments{
\item{snp1}{SNP index 1}

\item{genon}{Genotype matrix}

\item{depth_Ref}{Ref. allele depth}

\item{depth_Alt}{Alt. allele depth}

\item{freq}{Allele frequencies}

\item{LDmeasures}{LD measures (\in ("Dprime", "R2"))}
}
\description{
Title
}
