% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{readBCF_Contigs}
\alias{readBCF_Contigs}
\title{Reads contig names and lengths from BCF file as a data frame.}
\usage{
readBCF_Contigs(path)
}
\arguments{
\item{path}{Path to VCF or BCF file}
}
\description{
Reads contig names and lengths from BCF file as a data frame.
}
