% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{readBCF_GTandAD}
\alias{readBCF_GTandAD}
\title{Reads gentoype (0, 1, 2, NA coded) and allele depths from BCF. Optionally,
reads only from a specific region of the reference.}
\usage{
readBCF_GTandAD(path, region = NULL)
}
\arguments{
\item{path}{Path to VCF or BCF file}

\item{region}{A region (like 'Chr1:1-1000') to extract. `path` must be an
indexed BCF if this option is used.}
}
\description{
Reads gentoype (0, 1, 2, NA coded) and allele depths from BCF. Optionally,
reads only from a specific region of the reference.
}
