% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSLD-package.R
\docType{package}
\name{GUSLD-package}
\alias{GUSLD-package}
\alias{GUSLD}
\title{Genotyping Uncertainty with Sequencing data - Linkage Disequilibrium (GUSLD)}
\description{
\tabular{ll}{ Package: \tab GUSLD\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2018-10-22\cr License: \tab GPL 3\cr }
}
\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<timothy.bilton@agresearch.co.nz>
}
\keyword{package}
