% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSLD.R
\name{GUSLD}
\alias{GUSLD}
\title{Compute pairwise LD estimates using GUSLD}
\usage{
GUSLD(URobj, SNPpairs = NULL, indset = NULL, nClust = 3,
  LDmeasure = "r2", file = NULL, dp = 4)
}
\arguments{
\item{URobj}{An object of class UR created by the \code{\link{makeUR}} function.}

\item{SNPpairs}{Matrix of specifying the pairs of SNPs in which to calculate pairwise LD.}

\item{indset}{Vector of integer indices corresponding to individuals to be used in the LD calculations.
Used to specify a subset of individuals of the data for which LD is calculated.}

\item{nClust}{Integer number specifying the number of cores to use in the paralellization.}

\item{LDmeasure}{Character vector specifying which LD measures to calculate. Predefined
measures are \code{\link{Dcoef}}, \code{\link{Dprime}} and \code{\link{r2}} (see details below).
One can also specify their own LD measure (see examples).}

\item{file}{Character string giving the name of the file to write the LD results to. If NULL,
the function returns the LD results instead of writing them to a file.}

\item{dp}{Integer number specifying the number of decimal places to round the LD results.
Note: only works when \code{file} is not NULL.}
}
\description{
Function computes the LD estimates for high-throughput sequencing data using the methodology
by \insertCite{bilton2018genetics2;textual}{GUSbase}.
}
\examples{
## Define LD measure as the correlation coefficient
LD_r <- function(pA1,pA2,D){
return( D/sqrt((prod(c(pA1,pA2,1-c(pA1,pA2))))) )
}


}
\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
\author{
Timothy P. Bilton
}
