% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSLD.R
\name{GUSLD}
\alias{GUSLD}
\title{Compute pairwise LD estimates using GUSLD}
\usage{
GUSLD(URobj, SNPpairs = NULL, indsubset = NULL, nClust = 2,
  LDmeasure = "r2", filename = NULL, dp = 4)
}
\arguments{
\item{URobj}{An object of class UR created by the \code{\link{makeUR}} function.}

\item{SNPpairs}{Matrix specifying the pairs of SNPs in which to calculate pairwise LD. See below for details.}

\item{indsubset}{Vector of integer indices corresponding to individuals to be used in the LD calculations.
Used to specify a subset of individuals of the data for which LD is calculated.}

\item{nClust}{Integer number specifying the number of cores to use in the paralellization.}

\item{LDmeasure}{Character vector specifying which LD measures to calculate. Predefined
measures are \code{\link{Dcoef}}, \code{\link{Dprime}} and \code{\link{r2}} (see details below).
One can also specify their own LD measure (see examples).}

\item{filename}{Character string giving the name of the file to write the LD results to. If NULL,
the function returns the LD results instead of writing them to a file.}

\item{dp}{Integer number specifying the number of decimal places to round the LD results.
Note: only works when \code{file} is not NULL.}
}
\value{
A matrix object with named columns is returned where the columns are:
\itemize{
\item [\emph{LDmeasure}]: The LD estmate for the measure [\emph{LDmeasure}]. Note: if multiple LD measures have been specified,
then there will be a column for each LD measure
\item CHROM_SNP1: The chromosome number of the first SNP
\item CHROM_SNP2: The chromosome number of the second SNP
\item POS_SNP1: The position (in base pairs) of the first SNP on the chromosome
\item POS_SNP2: The position (in base pairs) of the second SNP on the chromosome
\item FREQ_SNP1: The allele frequency estimate of the first SNP (as used in the calculation of the LD measure)
\item FREQ_SNP2: The allele frequency estimate of the second SNP (as used in the calculation of the LD measure)
\item ERR_SNP1: The sequencing error estimate of the first SNP (as used in the calculation of the LD measure)
\item ERR_SNP2: The sequencing error estimate of the second SNP (as used in the calculation of the LD measure)
}
If \code{filename=NULL} then the matrix object is returned to the R workspace, otherwise if \code{filename} is specified
and corresponds to a vaild name, then the matrix object is written to a file called \emph{filename_GUSMap.txt}.
}
\description{
Function computes the LD estimates for high-throughput sequencing data using the methodology
by \insertCite{bilton2018genetics2;textual}{GUSbase}.
}
\details{
Argument \code{SNPpairs} must be an integer matrix with two columns, where the first column specify the
indices of the SNPs and the rows correspond to each SNP pair for which LD is to be computed.
If \code{SNPpairs=NULL}, then LD is estimated for all the SNP paris in the
dataset.

To reduce computation time, the calculations are performed in parallel using the \code{\link[foreach]{foreach}}
function. The \code{nClust} argument specifies the number of cores to use in the parallelization. Note: do not
set this argument to more than the number of cores available (or bad things can happen!).
}
\examples{

## Read in the deer data that accompanies the package
deerfile <- deerData()
rafile <- VCFtoRA(deerfile)
deer <- readRA(rafile)

## Create unrelated population
ur <- makeUR(deer)

###### LD estimation #######
## Estimate all the pairwise LD
LDres <- GUSLD(ur)
head(LDres)

## write results to file
GUSLD(ur, file="results")

## Specifying SNP pairs explicitly
# block vs block
pairs <- as.matrix(expand.grid(1:10,11:20))
LDres <- GUSLD(ur, SNPpairs=pairs)
head(LDres)

# Five SNPs before and after
temp <- rep(1:37,rep(5,37))
pairs <- cbind(temp, temp + 1:5)
pairs <- pairs[-which(pairs[,2] > 38),]
LDres <- GUSLD(ur, SNPpairs=pairs)
head(LDres)

##### Non-standard LD measure: #######
## Define LD measure as the correlation coefficient
LD_r <- function(pA1,pA2,D){
return( D/sqrt((prod(c(pA1,pA2,1-c(pA1,pA2))))) )
}
LDres <- GUSLD(ur, LDmeasure="LD_r")
head(LDres)

}
\references{
\insertRef{bilton2018genetics2}{GUSbase}
}
\author{
Timothy P. Bilton
}
